// src/app/main/distribution/constant/distribution.enums.ts
export enum DISTRIBUTION {
    COMPANY = "api/referencedata/company",
    GENDER = 'gender',
    CATEGORY = 'category',
    BRAND_TYPE = 'brandType',
    SUB_CATEGORY = 'subCategory',
    AG = 'ag',
    BRAND_NAME = 'brandName',
    PERIOD = 'api/referencedata/period'
}

export enum DistributionPath {
    DISTRIBUTIONAPIPATH = 'api/limits/distribution',
    SHOWROOMAPIPATH = 'api/limits/showroom',
    HEADERS = 'api/limits/header/distribution',
    SHOWROOMHEADERS = 'api/limits/header/showroom',
    DOWNLOAD = 'api/limits/download/distribution',
    SHOWROOMDOWNLOAD = 'api/limits/download/showroom'
}